requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
joblib
pillow
matplotlib
